# docs/swagger.yaml
basePath: /
definitions:
...
  permission.CreateInput:
    properties:
      action_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      resource_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
    type: object
  permission.ListResponse:
    properties:
      permissions:
        items:
          $ref: '#/definitions/permission.Permission'
        type: array
    type: object
  permission.Permission:
    properties:
      action_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      resource_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      role_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      updated_at:
        type: string
    type: object
  resource.CreateInput:
    properties:
      metadata:
        type: object
      name:
        example: documents
        type: string
    type: object
  resource.ListResponse:
    properties:
      resources:
        items:
          $ref: '#/definitions/resource.Resource'
        type: array
    type: object
  resource.Resource:
    properties:
      app_id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        example: 01912345-6789-7abc-def0-123456789abc
        type: string
      metadata:
        type: object
      name:
        example: documents
        type: string
      updated_at:
        type: string
    type: object
  resource.UpdateInput:
    properties:
      metadata:
        type: object
      name:
        example: documents
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    name: OpenAura
  description: |-
    Open-source user, role, and access management service.
    API version is selected via the X-API-Version header (not the URL path).
    App routes require an app API key (X-API-Key). Admin routes live under /admin and require an admin API key.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: OpenAura API
  version: "1.0"
paths:
...
  /resources:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resource.ListResponse'
      summary: List resources
      tags:
      - resources
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/resource.Resource'
      summary: Create resource
      tags:
      - resources
  /resources/{id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete resource
      tags:
      - resources
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resource.Resource'
      summary: Get resource
      tags:
      - resources
    patch:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Resource ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/resource.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resource.Resource'
      summary: Update resource
      tags:
      - resources
...
  /roles/{id}/permissions:
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by resource ID
        in: query
        name: resource_id
        type: string
      - description: Filter by action ID
        in: query
        name: action_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.ListResponse'
      summary: List role permissions
      tags:
      - permissions
    post:
      consumes:
      - application/json
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/permission.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/permission.Permission'
      summary: Create role permission
      tags:
      - permissions
  /roles/{id}/permissions/{permission_id}:
    delete:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission ID
        in: path
        name: permission_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      summary: Delete role permission
      tags:
      - permissions
    get:
      parameters:
      - default: "1"
        description: API version
        in: header
        name: X-API-Version
        required: true
        type: string
      - description: App API key
        in: header
        name: X-API-Key
        required: true
        type: string
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Permission ID
        in: path
        name: permission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/permission.Permission'
      summary: Get role permission
      tags:
      - permissions
...
schemes:
- http
swagger: "2.0"
...